This patch adds octet and IP address support
Some objects (notably MAC addresses and PortLists) require
the ability to send octets and this support was missing.
Octets are sent simply by sending a space separated string
with hex strings.  For example, the MAC address 01:00:1f:0b:00:0b
would be sent as a string "01 00 1f 0b 00 0b" with the add_oct()
method.